requirements.txt
numpy==1.21.3
tqdm==4.62.3
Pillow==8.3.2